model: Introduce a new copyright holder field to data model
In German law, the author and the copyright holder can be two separate
legal entities and therefore also need to be treated separately.

Introduce a new copyright holder field that is now the primary source
for copyright holder information. Authors are still only used as
copyright holders if the `addAuthorsToCopyrights` option is enabled.

For now, all package manager implementations set empty copyright
holders. Filling the copyright holder field is left as an exercise for
future actions. Right now, the only way to add copyright holders is via
curations.

This change resolves #4519.

Signed-off-by: Rainer Bieniek <extern.rainer.bieniek@porsche.de>
Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>